fix(sync): widen DHCP lease iaid to BIGINT; roll back before status write
DHCPv6 IAID is an unsigned 32-bit value (RFC 8415). Values above
2147483647 overflowed the Postgres INTEGER column and aborted the DHCP
sync flush with NumericValueOutOfRange.

The sync error handler then called _set_status() on the same session
whose transaction was already doomed, raising PendingRollbackError and
masking the original error. _set_status() now rolls back first.

- models/cache.py: iaid -> BigInteger
- alembic 0031: cache_dhcp_leases.iaid INTEGER -> BIGINT

+++ backend/app/models/cache.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Column, Integer, String, Boolean, DateTime
+from sqlalchemy import Column, Integer, BigInteger, String, Boolean, DateTime
 from app.database import Base
 
 
@@ -44,7 +44,7 @@ class CachedDHCPLease(Base):
     ip_address  = Column(String, nullable=False)
     mac_address = Column(String, default="")
     client_duid = Column(String, default="")
-    iaid        = Column(Integer, default=0)
+    iaid        = Column(BigInteger, default=0)  # DHCPv6 IAID is uint32 (RFC 8415)
     name        = Column(String, default="")
     description = Column(String, default="")
     source      = Column(String, nullable=False)
